+++ flutter/basic_client/lib/main.dart
@@ -1,3 +1,5 @@
+import 'package:basic_client/pages/login_form.dart';
+import 'package:basic_client/pages/source_list_page.dart';
 import 'package:flutter/material.dart';
 import 'package:english_words/english_words.dart';
 import 'package:provider/provider.dart';
@@ -21,7 +23,7 @@ class MyApp extends StatelessWidget {
         title: 'SD Flutter Client',
         theme: ThemeData(
           colorScheme: ColorScheme.fromSeed(
-            seedColor: const Color.fromARGB(255, 64, 83, 146),
+            seedColor: const Color.fromARGB(255, 43, 100, 255),
           ),
         ),
         home: const SDMain(title: 'SD Flutter test'),
@@ -46,6 +48,8 @@ class ProxyRequest {
   final String body;
   final Map<String, dynamic>? headers;
 
+  final jsonEncoder = JsonEncoder();
+
   Map<String, dynamic> toJson() => {
     'method': method,
     'path_query': pathQuery,
@@ -53,11 +57,17 @@ class ProxyRequest {
     'body': body,
     'headers': headers,
   };
+
+  String payload() {
+    final jsonEncoder = JsonEncoder();
+    return jsonEncoder.convert(toJson());
+  }
 }
 
 class MyAppState extends ChangeNotifier {
   // global application state lives here - we'd likely use something like Bloc for state management in a "real" app
   String authToken = '';
+  String journalistName = '';
   // gonna ignore token expiry for now but we'd need to keep an eye on it otherwise
   // authExpiry = DateTime.now()
   final sourceList = <String>[];
@@ -80,10 +90,6 @@ class MyAppState extends ChangeNotifier {
       headers: headers,
     );
 
-    final String payload = jsonEncoder.convert(
-      proxyRequest,
-    ); //this could probably be a proxyRequest class method
-
     if (proxyRequest.stream) {
       print("Streaming mode not implemented yet");
       return "";
@@ -100,22 +106,17 @@ class MyAppState extends ChangeNotifier {
         environment: originEnv,
       );
 
-      proc.stdin.writeln(payload);
-      final List<String> _lOut = [];
-      final List<String> _lErr =
-          []; // not actually using stderr for anything but let's grab it anyway
+      proc.stdin.writeln(proxyRequest.payload());
+      final List<String> lOut = [];
 
-      await proc.stdout.transform(utf8.decoder).forEach(_lOut.add);
-      await proc.stderr.transform(utf8.decoder).forEach(_lErr.add);
+      await proc.stdout.transform(utf8.decoder).forEach(lOut.add);
 
-      final _sOut = _lOut.join('\n');
-      final _sErr = _lErr.join('\n');
+      final sOut = lOut.join('\n');
 
       var theCode = await proc.exitCode;
       print("Exit Code is $theCode");
-      return _sOut;
+      return sOut;
     }
-    return ""; // lame, should throw an exception or something if we get this far
   }
 
   void syncSources() {
@@ -143,6 +144,7 @@ class MyAppState extends ChangeNotifier {
       if (fullJson['status'] == 200) {
         print("Token is ${bodyJson['token']}");
         authToken = bodyJson['token'];
+        journalistName = username;
         notifyListeners();
       } else {
         print("Error is ${bodyJson["message"]}");
@@ -192,201 +194,6 @@ class _SDMainState extends State<SDMain> {
       body: Container(
         color: Theme.of(context).colorScheme.primaryContainer,
         child: page,
-      ),
-    );
-  }
-}
-
-// Let's add the login form page
-class LoginForm extends StatefulWidget {
-  const LoginForm({super.key});
-
-  @override
-  LoginFormState createState() {
-    return LoginFormState();
-  }
-}
-
-class LoginFormState extends State<LoginForm> {
-  final _formKey = GlobalKey<FormState>();
-
-  final usernameController = TextEditingController();
-  final passphraseController = TextEditingController();
-  final oneTimeCodeController = TextEditingController();
-
-  // I find this to be lame
-  @override
-  void dispose() {
-    usernameController.dispose();
-    passphraseController.dispose();
-    oneTimeCodeController.dispose();
-    super.dispose();
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    usernameController.text = "journalist";
-    passphraseController.text =
-        "correct horse battery staple profanity oil chewy";
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var appState = context.watch<MyAppState>();
-
-    return Form(
-      key: _formKey,
-      child: Column(
-        children: <Widget>[
-          Text("Log in here: "),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: TextFormField(
-              controller: usernameController,
-              decoration: const InputDecoration(
-                border: OutlineInputBorder(),
-                labelText: 'Username:',
-              ),
-              // The validator receives the text that the user has entered.
-              validator: (value) {
-                if (value == null || value.isEmpty) {
-                  return 'Please enter your username';
-                }
-                return null;
-              },
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: TextFormField(
-              controller: passphraseController,
-              decoration: const InputDecoration(
-                border: OutlineInputBorder(),
-                labelText: 'Passphrase:',
-              ),
-              // The validator receives the text that the user has entered.
-              validator: (value) {
-                if (value == null || value.isEmpty) {
-                  return 'Please enter your passphrase';
-                }
-                return null;
-              },
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: TextFormField(
-              controller: oneTimeCodeController,
-              decoration: const InputDecoration(
-                border: OutlineInputBorder(),
-                labelText: 'Two-factor token:',
-              ),
-              // The validator receives the text that the user has entered.
-              validator: (value) {
-                if (value == null || value.isEmpty || value.length != 6) {
-                  return 'Please enter a valid 6-char 2FA token';
-                }
-                return null;
-              },
-            ),
-          ),
-          ElevatedButton(
-            onPressed: () {
-              // Validate returns true if the form is valid, or false otherwise.
-              if (_formKey.currentState!.validate()) {
-                ScaffoldMessenger.of(
-                  context,
-                ).showSnackBar(const SnackBar(content: Text('Logging in')));
-                appState.tryLogin(
-                  usernameController.text,
-                  passphraseController.text,
-                  oneTimeCodeController.text,
-                );
-              }
-            },
-            child: const Text('Submit'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// Let's add the source list page...
-// Define the Source List view
-class SourceListView extends StatelessWidget {
-  const SourceListView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    var appstate = context.watch<MyAppState>();
-
-    return Container(
-      width: 350,
-      color: const Color.fromARGB(255, 154, 195, 241),
-      child: ListView.separated(
-        scrollDirection: Axis.vertical,
-        padding: const EdgeInsets.all(8),
-        itemCount: appstate.sourceList.length,
-        itemBuilder: (BuildContext context, int index) {
-          return SizedBox(height: 50, child: Text(appstate.sourceList[index]));
-        },
-        separatorBuilder: (BuildContext context, int index) => const Divider(),
-      ),
-    );
-  }
-}
-
-class SourceListPage extends StatelessWidget {
-  const SourceListPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    var appstate = context.watch<MyAppState>();
-    String sourceText = "No Sources";
-    if (appstate.sourceList.isNotEmpty) {
-      sourceText = "${appstate.sourceList.length} Sources";
-    }
-
-    return Scaffold(
-      body: Column(
-        children: [
-          Row(
-            mainAxisSize: MainAxisSize.max,
-            mainAxisAlignment: MainAxisAlignment.start,
-            children: [
-              Padding(
-                padding: const EdgeInsets.all(10.0),
-                child: Text(sourceText),
-              ),
-              ElevatedButton.icon(
-                onPressed: () {
-                  appstate.syncSources();
-                },
-                icon: Icon(Icons.sync),
-                label: Text('Sync'),
-              ),
-              SizedBox(width: 10),
-              ElevatedButton(
-                onPressed: () {
-                  appstate.logout();
-                },
-                child: Text('Logout'),
-              ),
-            ],
-          ),
-          Expanded(
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: [
-                SourceListView(),
-                Expanded(child: Placeholder()),
-              ],
-            ),
-          ),
-        ],
       ),
     );
   }
